+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-02 16:48:47
-// Commit: fix(core): fix memory leak by cancelling core subscription
+// Updated at: 2026-04-02 19:40:08
+// Commit: feat(core): implement responsive core view component
 
 class coreComponentState {
-  final String commitRef = '69';
-  final String lastMessage = 'fix(core): fix memory leak by cancelling core subscription';
-  final DateTime timestamp = DateTime.parse('2026-04-02T16:48:47+05:00');
+  final String commitRef = '79';
+  final String lastMessage = 'feat(core): implement responsive core view component';
+  final DateTime timestamp = DateTime.parse('2026-04-02T19:40:08+05:00');
 }